+++ lib/custom_widgets/go_anonymous_button.dart
@@ -1,13 +1,24 @@
+import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/material.dart';
 class SignInAnonymously extends StatelessWidget {
+  Future<void> signInAnonymously() async {
+    try{
+      final userCredential = await FirebaseAuth.instance.signInAnonymously();
+      print("${userCredential.user.uid}");
+    }
+   catch(e){
+      print(e.toString());
+   }
+  }
+
   @override
   Widget build(BuildContext context) {
-    return ElevatedButton(onPressed: (){},
+    return ElevatedButton(onPressed: signInAnonymously,
       style: ButtonStyle(
         elevation: MaterialStateProperty.all(6),
         shape: MaterialStateProperty.all(StadiumBorder()),
         padding: MaterialStateProperty.all(EdgeInsets.all(14)),
-        backgroundColor: MaterialStateProperty.all(Colors.white),
+        backgroundColor: MaterialStateProperty.all(Colors.yellow[50]),
       ),
       child: Row(
         mainAxisAlignment: MainAxisAlignment.center,
