+++ lib/main.dart
@@ -1,16 +1,15 @@
-
+import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
-import 'package:time_tracker/custom_widgets/facebook_SignIn_button.dart';
-import 'package:time_tracker/custom_widgets/go_anonymous_button.dart';
-import 'package:time_tracker/custom_widgets/google_SignIn_button.dart';
-import 'package:time_tracker/custom_widgets/SignIn_with_email_button.dart';
+import 'package:time_tracker/pages/landing_page.dart';
+import 'package:time_tracker/pages/sign_in_page.dart';
 
-void main() {
+void main() async {
+  WidgetsFlutterBinding.ensureInitialized();
+  await Firebase.initializeApp();
   runApp(MyApp());
 }
 
 class MyApp extends StatelessWidget {
-  // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
@@ -18,61 +17,8 @@ class MyApp extends StatelessWidget {
       theme: ThemeData(
         primarySwatch: Colors.indigo,
       ),
-      home: MySigninPage(),
+      home: LandingPage(),
     );
   }
 }
-class MySigninPage extends StatelessWidget {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        centerTitle: true,
-        title: Text(
-          "Welcome",
-        ),
-      ),
-      body: Container(
-        padding: EdgeInsets.all(16),
-        height: MediaQuery.of(context).size.height,
-        width: MediaQuery.of(context).size.width,
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: [
-            Text(
-              "Sign In",
-              style: TextStyle(
-                fontSize: 30,
-                fontWeight: FontWeight.w700,
-              ),
-            ),
-            SizedBox(
-              height: 50,
-            ),
-            SignInGoogle(),
-            SizedBox(
-              height: 22,
-            ),
-            SignInFacebook(),
-            SizedBox(
-              height: 22,
-            ),
-            SignInEmail(),
-            SizedBox(
-              height: 22,
-            ),
-            Text(
-              "or"
-            ),
-            SizedBox(
-              height: 22,
-            ),
-            SignInAnonymously(),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
