+++ lib/pages/landing_page.dart
@@ -1,0 +1,13 @@
+import 'package:flutter/material.dart';
+import 'package:time_tracker/pages/sign_in_page.dart';
+class LandingPage extends StatefulWidget {
+  @override
+  _LandingPageState createState() => _LandingPageState();
+}
+
+class _LandingPageState extends State<LandingPage> {
+  @override
+  Widget build(BuildContext context) {
+    return MySignInPage();
+  }
+}
