+++ pubspec.yaml
@@ -23,6 +23,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  firebase_core: ^1.0.1
+  firebase_auth: ^1.0.1
 
 
   # The following adds the Cupertino Icons font to your application.
